backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.3
beautifulsoup4==4.13.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
numpy==2.2.3
openai==0.28.1
outcome==1.3.0
passlib==1.7.4
Pillow==10.1.0
pyarrow==19.0.0
pycparser==2.21
PyJWT==2.8.0
pylance==0.23.0
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.2
